Cleanup. Stack size problems

+++ sn/sni_logbuffer.cpp
@@ -91,8 +91,8 @@ namespace SNI
 	void SNI_LogBuffer::DerivationJS(ostream & p_Stream, long p_MaxLogEntries)
 	{
 		p_Stream << "{\"derivationhtml\": \"";
-		size_t actualDepth = SNI_Thread::GetThread()->GetFrameStackDepth();
-		size_t minDepth = ULONG_MAX;
+		long actualDepth = SNI_Thread::GetThread()->GetFrameStackDepth();
+		long minDepth = LONG_MAX;
 		long entries = 0;
 		m_Mutex.lock();
 		for (auto it = m_Buffer.rbegin(); it != m_Buffer.rend() && (p_MaxLogEntries <= 0 || entries < p_MaxLogEntries); it++, entries++)
@@ -107,7 +107,7 @@ namespace SNI
 		vector<SNI_LogLine *> stackLastLines;
 		for (auto it = m_Buffer.rbegin(); it != m_Buffer.rend() && ((p_MaxLogEntries <= 0 || entries < p_MaxLogEntries) || currentDepth > minDepth); it++, entries++)
 		{
-			for (size_t j = currentDepth; j < it->m_Depth; j++)
+			for (long j = currentDepth; j < it->m_Depth; j++)
 			{
 				ASSERTM(j - minDepth + 1 == stackLastLines.size(), "Internal error 1");
 				stackLastLines.push_back(NULL);
@@ -121,7 +121,7 @@ namespace SNI
 				}
 			}
 			ASSERTM(it->m_Depth - minDepth + 1 <= stackLastLines.size(), "Internal error 2");
-			for (size_t j = currentDepth; it->m_Depth < j; j--)
+			for (long j = currentDepth; it->m_Depth < j; j--)
 			{
 				ASSERTM(j - minDepth + 1 == stackLastLines.size(), "Internal error 3");
 				SNI_LogLine *line = stackLastLines.back();
@@ -153,7 +153,7 @@ namespace SNI
 				p_Stream << "<p>" << it->m_StepCount << ": " << EscapeStringToJSON(html) << "</p>\\n";
 			}
 		}
-		for (size_t j = currentDepth; minDepth <= j; j--)
+		for (long j = currentDepth; minDepth <= j; j--)
 		{
 			ASSERTM(j - minDepth + 1 == stackLastLines.size(), "Internal error 5");
 			SNI_LogLine *line = stackLastLines.back();
